feature[test]: refactoring issue #158
