+++ mockk/common/src/test/kotlin/io/mockk/it/MockkClassTest.kt
@@ -1,16 +1,13 @@
 package io.mockk.it
 
 import io.mockk.every
+import io.mockk.mockk
 import io.mockk.mockkClass
 import io.mockk.verify
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
 class MockkClassTest {
-
-    class MockCls {
-        fun op(a: Int, b: Int) = a + b
-    }
 
     val mock = mockkClass(MockCls::class)
 
@@ -24,5 +21,27 @@ class MockkClassTest {
         assertEquals(5, mock.op(3, 4))
 
         verify { mock.op(3, 4) }
+    }
+
+    private val targetClass = mockk<TestClass>()
+
+    /**
+     * See issue #158
+     */
+    @Test
+    fun testNothingIsNotThrowingNPE() {
+        every { targetClass.alwaysThrows() } answers {
+            throw IllegalArgumentException("this is a test")
+        }
+    }
+
+    class TestClass {
+        fun alwaysThrows() : Nothing {
+            throw RuntimeException("this can be any exception")
+        }
+    }
+
+    class MockCls {
+        fun op(a: Int, b: Int) = a + b
     }
 }
